pyproject.toml
[tool.poetry]
name = "ppr-project"
version = "1.4.1"
description = "PPR project site"
authors = ["Ivan Skvortsov <pprofcheg@gmail.com>"]
license = "LICENSE"
readme = "README.md"
packages = [{include = "ppr_project"}]

[tool.poetry.dependencies]
python = "^3.8"
celery = "^5.2.7"
Django = "^4.1.1"
django-simple-history = "^3.1.1"
gunicorn = "^20.1.0"
openpyxl = "^3.0.10"
Pillow = "^9.2.0"
psycopg2-binary = "^2.9.3"
redis = "^4.3.4"
sentry-sdk = "^1.9.8"
sorl-thumbnail = "^12.9.0"
tqdm = "^4.64.1"
python-dotenv = "^0.21.0"
djangorestframework = "^3.14.0"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^3.6.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
